pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "stylepro"
dynamic = ["version"]
description = "A visual style editor overlay for Streamlit, Dash, and more"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["streamlit", "dash", "theming", "styling", "ui", "css"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: User Interfaces",
]
dependencies = []  # zero hard requirements; all framework support is opt-in

[project.optional-dependencies]
streamlit = ["streamlit>=1.30"]
dash      = ["dash>=2.0"]
yaml      = ["PyYAML>=6.0"]
dev       = ["pytest>=7", "pytest-asyncio>=0.23", "httpx>=0.24"]
all       = ["streamlit>=1.30", "dash>=2.0", "PyYAML>=6.0"]

[tool.hatch.version]
path = "stylepro/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["stylepro"]
artifacts = ["stylepro/editor/static/**"]

[tool.hatch.build.targets.sdist]
include = [
    "stylepro/",
    "tests/",
    "examples/",
    "pyproject.toml",
    "MANIFEST.in",
    "README.md",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
